Create test program for C_c handling
Move the signal handling code to a separate C++ file and create a very
simple program to allow pressing C_c/CTRL+C multiple times.
